git lfs: Add support for ignoring files managed by git-lfs
This PR adds support for ignoring files handled by git lfs. jj previously detected these files as changed as it expected the original file content before the git lfs checkout to be present. With this change it's possible to opt in (via the git.ignore-lfs-files setting) into just ignoring this files from jj. They then need to be manually handled via e.g. git if they are changed.

From a technical point of view this change parses all .gitattributes files in the current repository and constructs a ignore list from that. This ignore list is then applied through all subcommands.

co-authored-by: Brandon Sprague <brandon@sprague.mx>